LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-button-padding.html

<!DOCTYPE html>
<script src="../../../resources/js-test-pre.js"></script>
<body>
<video src="../../content/test.mp4" style="width: 320px; height: 240px;" controls autoplay></video>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing that buttons on iOS have extra padding and remain centered.");

const mediaControls = window.internals.shadowRoot(document.querySelector("video")).lastElementChild;

shouldBecomeEqual("mediaControls.querySelector('button.fullscreen') instanceof HTMLButtonElement", "true", () => {
    shouldBeEqualToString("getComputedStyle(mediaControls.querySelector('button.fullscreen')).padding", "8px");
    shouldBeEqualToString("getComputedStyle(mediaControls.querySelector('button.fullscreen')).transform", "matrix(1, 0, 0, 1, -8, 0)");
    finishJSTest();
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
