LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-button-padding-expected.txt

Testing that buttons on iOS have extra padding and remain centered.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.querySelector('button.fullscreen') instanceof HTMLButtonElement became true
PASS getComputedStyle(mediaControls.querySelector('button.fullscreen')).padding is "8px"
PASS getComputedStyle(mediaControls.querySelector('button.fullscreen')).transform is "matrix(1, 0, 0, 1, -8, 0)"
PASS successfullyParsed is true

TEST COMPLETE
